+++ easy_post/lib/src/screens/post.detail.screen.dart
@@ -1,3 +1,4 @@
+import 'package:easy_comment/easy_comment.dart';
 import 'package:easy_post_v2/easy_post_v2.dart';
 import 'package:easyuser/easyuser.dart';
 import 'package:flutter/material.dart';
@@ -21,19 +22,40 @@ class _PostDetailScreenState extends State<PostDetailScreen> {
       appBar: AppBar(
         title: const Text('PostDetail'),
       ),
-      body: Column(
-        children: [
-          UserDoc(
-            uid: widget.post.uid,
-            builder: (user) => user == null
-                ? const SizedBox.shrink()
-                : UserAvatar(
-                    user: user,
+      body: CustomScrollView(
+        slivers: [
+          SliverPadding(
+            padding: const EdgeInsets.all(16),
+            sliver: SliverToBoxAdapter(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  UserDoc(
+                    uid: widget.post.uid,
+                    builder: (user) => user == null
+                        ? const SizedBox.shrink()
+                        : UserAvatar(
+                            user: user,
+                          ),
                   ),
+                  Text(widget.post.title),
+                  Text(widget.post.content),
+                ],
+              ),
+            ),
           ),
-          Text(widget.post.title),
-          Text(widget.post.content),
+          CommentListTreeView(
+            documentReference: widget.post.ref,
+          ),
         ],
+      ),
+      bottomNavigationBar: SafeArea(
+        child: CommentFakeInputBox(onTap: () {
+          CommentService.instance.showCommentEditDialog(
+            context: context,
+            documentReference: widget.post.ref,
+          );
+        }),
       ),
     );
   }