+++ easy_comment/CHANGELOG.md
@@ -1,5 +1,8 @@
+## 0.0.3
+* Add. CommentListTreeView for tree style vertical line comment listing.
+
 ## 0.0.2
-* initial release.
+* Add. widgets for comment listing.
 
 ## 0.0.1
 * initial release.